+++ .storybook/preview.tsx
@@ -13,3 +13,9 @@ const withMagma = (StoryFn: Function, context: StoryContext) => {
 };
 
 export const decorators = [withMagma, withPerformance];
+
+// https://storybook.js.org/docs/react/writing-stories/parameters#global-parameters
+export const parameters = {
+  // https://storybook.js.org/docs/react/essentials/actions#automatically-matching-args
+  actions: { argTypesRegex: '^on.*' },
+};