+++ packages/react-magma-dom/src/components/Button/index.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import { StyledButton } from '../StyledButton';
+import { useIsInverse } from '../../inverse';
 
 export enum ButtonVariant {
   solid = 'solid', //default
@@ -96,8 +97,10 @@ export const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
   (props, ref) => {
     const { children, ...rest } = props;
 
+    const isInverse = useIsInverse(props.isInverse);
+
     return (
-      <StyledButton {...rest} ref={ref}>
+      <StyledButton {...rest} isInverse={isInverse} ref={ref}>
         {children}
       </StyledButton>
     );