+++ packages/react-magma-dom/src/components/IconButton/index.tsx
@@ -12,6 +12,7 @@ import {
 import { IconProps } from 'react-magma-icons';
 import { omit, Omit, XOR } from '../../utils';
 import { ThemeContext } from '../../theme/ThemeContext';
+import { useIsInverse } from '../../inverse';
 
 export enum ButtonIconPosition {
   left = 'left',
@@ -101,6 +102,8 @@ export const IconButton = React.forwardRef<HTMLButtonElement, IconButtonProps>(
 
     const other = omit(['iconPosition', 'textPosition'], rest);
 
+    const isInverse = useIsInverse(props.isInverse);
+
     if (icon && !children) {
       return (
         <StyledButton
@@ -108,6 +111,7 @@ export const IconButton = React.forwardRef<HTMLButtonElement, IconButtonProps>(
           ref={ref}
           color={color ? color : ButtonColor.primary}
           iconOnly
+          isInverse={isInverse}
           shape={shape ? shape : ButtonShape.round}
           size={size ? size : ButtonSize.medium}
           variant={variant ? variant : ButtonVariant.solid}
@@ -127,6 +131,7 @@ export const IconButton = React.forwardRef<HTMLButtonElement, IconButtonProps>(
         {...other}
         ref={ref}
         color={color ? color : ButtonColor.primary}
+        isInverse={isInverse}
         shape={shape ? shape : ButtonShape.fill}
         size={size ? size : ButtonSize.medium}
         textTransform={