+++ packages/react-magma-dom/src/components/Table/TablePagination.tsx
@@ -8,6 +8,7 @@ import styled from '@emotion/styled';
 import { ThemeContext } from '../../theme/ThemeContext';
 import { ThemeInterface } from '../../theme/magma';
 import { I18nContext } from '../../i18n';
+import { useIsInverse } from '../../inverse';
 
 export interface TablePaginationProps
   extends React.HTMLAttributes<HTMLDivElement> {
@@ -65,7 +66,6 @@ export const TablePagination = React.forwardRef<
   const {
     testId,
     count,
-    isInverse,
     onChangePage,
     onChangeRowsPerPage,
     page = 0,
@@ -75,6 +75,8 @@ export const TablePagination = React.forwardRef<
 
   const theme = React.useContext(ThemeContext);
   const i18n = React.useContext(I18nContext);
+
+  const isInverse = useIsInverse(props.isInverse);
 
   const selectContainerStyle = { display: 'flex', alignItems: 'center' };
   const isLastPage = (page + 1) * rowsPerPage >= count;